+++ docs/changelog.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 Each revision is versioned by the date of the revision.
 
+## 2026-08-26
+
+- Added `get-loadbalancer-ip` Juju action to expose the external LoadBalancer address.
+
 ## 2026-08-25
 
 - Fixed several documentation issues.